backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Task Queue
celery==5.3.4
redis==5.0.1

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# ML Core
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.2.0
joblib==1.3.2

# Deep Learning (Optional)
tensorflow==2.15.0
# torch==2.1.2

# Explainability
shap==0.44.0
lime==0.2.0.1

# Drift Detection
river==0.21.0
scipy==1.11.4
seaborn==0.12.2
matplotlib==3.8.2

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Logging & Monitoring
structlog==24.1.0

# Cloud SDKs (Optional)
boto3==1.34.28
google-cloud-storage==2.14.0

# Docker
docker==7.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
